move the feed inside the release notes util
