+++ package.json
@@ -29,6 +29,7 @@
     "babel-plugin-import-glob-array": "0.2.0",
     "babel-plugin-transform-define": "2.0.0",
     "bootstrap": "4.5.3",
+    "feed": "^4.2.1",
     "glob": "^7.1.6",
     "gray-matter": "^4.0.2",
     "moment": "2.29.1",
@@ -41,6 +42,7 @@
     "react-hook-form": "6.9.6",
     "react-syntax-highlighter": "15.2.1",
     "react-moment": "1.0.0",
+    "react-syntax-highlighter": "15.2.1",
     "remark-external-links": "^8.0.0",
     "sass": "1.27.0",
     "sharp": "0.26.2",
